making feature extraction modular
