+++ src/napari_convpaint/conv_paint_utils.py
@@ -1,28 +1,20 @@
 from collections import OrderedDict
-from typing import Any
 import warnings
-from pathlib import Path
 
 import torch
 import numpy as np
 import skimage.transform
 import skimage.morphology as morph
 import pandas as pd
-from joblib import dump, load
 from joblib import Parallel, delayed
-import yaml
-import zarr
 import einops as ein
 
-import torchvision.models as models
 from torch import nn
 from sklearn.model_selection import train_test_split
 from sklearn.ensemble import RandomForestClassifier
 
-from torchvision import transforms
 from torch.nn.functional import interpolate as torch_interpolate
 
-from .conv_parameters import Param
 
 def get_device(use_cuda=None):
     if torch.cuda.is_available() and (use_cuda is None or use_cuda==True):
@@ -30,8 +22,6 @@ def get_device(use_cuda=None):
     else:
         device = torch.device('cpu')
     return device
-    
-
 
 def normalize_image(image, image_mean, image_std):
     """
@@ -227,143 +217,6 @@ def filter_image_multioutputs(image, hookmodel, scalings=[1], order=0,
 
     return all_scales
 
-
-def filter_image_multichannels(image, hookmodel, scalings=[1], order=0,
-                              image_downsample=1):
-    """Recover the outputs of chosen layers of a pytorch model. Layers and model are
-    specified in the hookmodel object. If image has multiple channels, each channel
-    is processed separately.
-    
-    Parameters
-    ----------
-    image : np.ndarray
-        2d Image to filter
-    hookmodel : Hookmodel
-        Model to extract features from
-    scalings : list of ints, optional
-        Downsampling factors, by default None
-    order : int, optional
-        Interpolation order for low scale resizing,
-        by default 0
-    image_downsample : int, optional
-        Downsample image by this factor before extracting features, by default 1
-
-    Returns
-    -------
-    all_scales : list of np.ndarray
-        List of filtered images. The number of images is C x Sum_i(F_i x S) where C is the number of channels,
-        F_i is the number of filters of the ith layer and S the number of scaling factors.
-        
-    """
-
-    input_channels = hookmodel.named_modules[0][1].in_channels
-    image = np.asarray(image, dtype=np.float32)
-    
-    if image.ndim == 2:
-        image = image[::image_downsample, ::image_downsample]
-        image = np.ones((input_channels, image.shape[0], image.shape[1]), dtype=np.float32) * image
-        image_series = [image]
-    elif image.ndim == 3:
-        image = image[:, ::image_downsample, ::image_downsample]
-        image_series = [np.ones((input_channels, im.shape[0], im.shape[1]), dtype=np.float32) * im for im in image]
-
-    int_mode = 'bilinear' if order > 0 else 'nearest'
-    align_corners = False if order > 0 else None
-
-    all_scales = []
-    with torch.no_grad():
-        for image in image_series:
-            for s in scalings:
-                im_tot = image[:, ::s, ::s]
-                im_torch = torch.tensor(im_tot[np.newaxis, ::])
-                hookmodel.outputs = []
-                try:
-                    _ = hookmodel(im_torch)
-                except AssertionError as ea:
-                    pass
-                except Exception as ex:
-                    raise ex
-
-                for im in hookmodel.outputs:
-                    if image.shape[1:3] != im.shape[2:4]:
-                        im = torch_interpolate(im, size=image.shape[1:3], mode=int_mode, align_corners=align_corners)
-                        '''out_np = skimage.transform.resize(
-                            image=out_np,
-                            output_shape=(1, out_np.shape[1], image.shape[1], image.shape[2]),
-                            preserve_range=True, order=order)'''
-
-                    out_np = im.cpu().detach().numpy()
-                    all_scales.append(out_np)
-
-    return all_scales
-
-
-def predict_image(image, model, classifier, scalings=[1], order=0,
-                  use_min_features=True, image_downsample=1,
-                  ):
-    """
-    Given a filter model and a classifier, predict the class of 
-    each pixel in an image.
-
-    Parameters
-    ----------
-    image: 2d array
-        image to segment
-    model: Hookmodel
-        model to extract features from
-    classifier: sklearn classifier
-        classifier to use for prediction
-    scalings: list of ints
-        downsampling factors
-    order: int
-        interpolation order for low scale resizing
-    use_min_features: bool
-        if True, use the minimum number of features per layer
-    image_downsample: int, optional
-        downsample image by this factor before extracting features, by default 1
-
-    Returns
-    -------
-    predicted_image: 2d array
-        predicted image with classes
-
-    """
-    
-    filter_fun = filter_image_multichannels
-
-    if use_min_features:
-        max_features = np.min(model.features_per_layer)
-        all_scales = filter_fun(
-            image, model, scalings=scalings, order=order,
-            image_downsample=image_downsample)
-        all_scales = [a[:, 0:max_features, :, :] for a in all_scales]
-        tot_filters = max_features * len(all_scales)
-
-    else:
-        # max_features = np.max(model.features_per_layer)
-        all_scales = filter_fun(
-            image, model, scalings=scalings, order=order,
-            image_downsample=image_downsample)
-        tot_filters = np.sum(a.shape[1] for a in all_scales)
-    
-    tot_filters = int(tot_filters)
-    rows = np.ceil(image.shape[-2] / image_downsample).astype(int)
-    cols = np.ceil(image.shape[-1] / image_downsample).astype(int)
-    all_pixels = pd.DataFrame(
-        np.reshape(np.concatenate(all_scales, axis=1), newshape=(tot_filters, rows * cols)).T)
-
-    predictions = classifier.predict(all_pixels)
-
-    predicted_image = np.reshape(predictions, [rows, cols])
-    if image_downsample > 1:
-        predicted_image = skimage.transform.resize(
-            image=predicted_image,
-            output_shape=(image.shape[-2], image.shape[-1]),
-            preserve_range=True, order=1).astype(np.uint8)
-
-    return predicted_image
-
-
 def parallel_predict_image(image, model, classifier, scalings=[1], order=0,
                   use_min_features=True, image_downsample=1, use_dask=False):
     """
@@ -454,7 +307,7 @@ def parallel_predict_image(image, model, classifier, scalings=[1], order=0,
 
             if use_dask:
                 processes.append(client.submit(
-                    predict_image, image_block, model, classifier,
+                    model.predict_image, image_block, model, classifier,
                     scalings, order, use_min_features, image_downsample))
                 
                 min_row_ind_collection.append(min_row_ind)
@@ -467,7 +320,7 @@ def parallel_predict_image(image, model, classifier, scalings=[1], order=0,
                 new_min_row_ind_collection.append(new_min_row_ind)
 
             else:
-                predicted_image = predict_image(image_block, model, classifier,
+                predicted_image = model.predict_image(image_block, model, classifier,
                     scalings, order, use_min_features, image_downsample)
                 crop_pred = predicted_image[
                     new_min_row_ind: new_max_row_ind,
@@ -488,101 +341,7 @@ def parallel_predict_image(image, model, classifier, scalings=[1], order=0,
     
     return predicted_image_complete
 
-
-def load_single_layer_vgg16(keep_rgb=False):
-    """Load VGG16 model from torchvision, keep first layer only
-    
-    Parameters
-    ----------
-
-    Returns
-    -------
-    model: torch model
-        model for pixel feature extraction
-    keep_rgb: bool
-        if True, keep model with three input channels, otherwise convert to single channel
-    
-    """
-
-    vgg16 = models.vgg16(weights=models.VGG16_Weights.IMAGENET1K_V1)
-    pretrained_dict = vgg16.state_dict()
-    
-    if keep_rgb:
-        model = nn.Sequential(OrderedDict([('conv1', nn.Conv2d(3, 64, 3, 1, 1))]))
-    else:
-        model = nn.Sequential(OrderedDict([('conv1', nn.Conv2d(1, 64, 3, 1, 1))]))
-
-    reduced_dict = model.state_dict()
-
-    if keep_rgb:
-        reduced_dict['conv1.weight'] = pretrained_dict['features.0.weight']
-        reduced_dict['conv1.bias'] = pretrained_dict['features.0.bias']
-    else:
-        reduced_dict['conv1.weight'][:, 0, :, :] = pretrained_dict['features.0.weight'][:, :, :, :].sum(axis=1)
-        reduced_dict['conv1.bias'] = pretrained_dict['features.0.bias']
-
-    model.load_state_dict(reduced_dict)
-
-    return model
-
-def get_multiscale_features(model, image, annotations, scalings, order=0,
-                            use_min_features=True,
-                            image_downsample=1):
-    """Given an image and a set of annotations, extract multiscale features
-    
-    Parameters
-    ----------
-    model : Hookmodel
-        Model to extract features from
-    image : np.ndarray
-        Image to extract features from (currently 2D only or 2D multichannel (C,H,W))
-    annotations : np.ndarray
-        Annotations (1,2) to extract features from (currently 2D only)
-    scalings : list of ints
-        Downsampling factors
-    order : int, optional
-        Interpolation order for low scale resizing, by default 0
-    use_min_features : bool, optional
-        Use minimal number of features, by default True
-    image_downsample : int, optional
-        Downsample image by this factor before extracting features, by default 1
-
-    Returns
-    -------
-    extracted_features : np.ndarray
-        Extracted features. Dimensions npixels x nfeatures * nbscales
-    """
-
-    filter_fun = filter_image_multichannels
-
-    if use_min_features:
-        max_features = np.min(model.features_per_layer)
-    else:
-        max_features = np.max(model.features_per_layer)
-    # test with minimal number of features i.e. taking only n first features
-    rows = np.ceil(image.shape[-2] / image_downsample).astype(int)
-    cols = np.ceil(image.shape[-1] / image_downsample).astype(int)
-    full_annotation = np.ones((max_features, rows, cols), dtype=np.bool_)
-    full_annotation = full_annotation * annotations[::image_downsample, ::image_downsample] > 0
-
-    all_scales = filter_fun(
-        image, model, scalings, order=order,
-        image_downsample=image_downsample)
-    if use_min_features:
-        all_scales = [a[:, 0:max_features, :, :] for a in all_scales]
-    all_values_scales = []
-
-    for ind, a in enumerate(all_scales):
-        n_features = a.shape[1]
-        extract = a[0, full_annotation[0:n_features]]
-
-        all_values_scales.append(np.reshape(extract, (n_features, int(extract.shape[0] / n_features))).T)
-    extracted_features = np.concatenate(all_values_scales, axis=1)
-
-    return extracted_features
-
-
-def get_features_current_layers(model, image, annotations, scalings=[1],
+def get_features_current_layers(image, annotations, model=None, scalings=[1],
                                 order=0, use_min_features=True,
                                 image_downsample=1):
     """Given a potentially multidimensional image and a set of annotations,
@@ -590,13 +349,13 @@ def get_features_current_layers(model, image, annotations, scalings=[1],
     
     Parameters
     ----------
-    model : Hookmodel
-        Model to extract features from
     image : np.ndarray
         2D or 3D Image to extract features from. If 3D but annotations are 2D,
         image is treated as multichannel and not image series
     annotations : np.ndarray
         2D, 3D Annotations (1,2) to extract features from
+    model : feature extraction model
+        Model to extract features from the image
     scalings : list of ints
         Downsampling factors
     order : int, optional
@@ -613,6 +372,9 @@ def get_features_current_layers(model, image, annotations, scalings=[1],
     targets : pandas Series
         Target values
     """
+
+    if model is None:
+        raise ValueError('Model must be provided')
 
     # get indices of first dimension of non-empty annotations. Gives t/z indices
     if annotations.ndim == 3:
@@ -656,8 +418,10 @@ def get_features_current_layers(model, image, annotations, scalings=[1],
                 boxes['bbox-1'][i]:boxes['bbox-3'][i]
             ]
 
-            extracted_features = get_multiscale_features(
-                model, im_crop, annot_crop, scalings,
+            extracted_features = model.get_features(
+                image=im_crop,
+                annotations=annot_crop,
+                scalings=scalings,
                 order=order, use_min_features=use_min_features,
                 image_downsample=image_downsample)
             all_values.append(extracted_features)
@@ -1185,238 +949,3 @@ def train_classifier(features, targets):
     random_forest.fit(X, y)
 
     return random_forest
-
-
-def load_trained_classifier(model_path):
-    model_path = Path(model_path)
-    random_forest = load(model_path)
-
-    param = Param()
-    with open(model_path.parent.joinpath('convpaint_params.yml')) as file:
-        documents = yaml.full_load(file)
-    for k in documents.keys():
-        setattr(param, k, documents[k])
-
-    return random_forest, param
-
-
-# encapsulate of pytorch model that includes hooks and outputs for certain layers
-class Hookmodel():
-    """Class to extract features from a pytorch model using hooks on chosen layers.
-
-    Parameters
-    ----------
-    model_name : str
-        Name of model to use. Currently only 'vgg16' and 'single_layer_vgg16',
-         'single_layer_vgg16_rgb' are supported.
-    model : torch model, optional
-        Model to extract features from, by default None
-    use_cuda : bool, optional
-        Use cuda, by default False
-    param : Param, optional
-        Parameters for model, by default None
-        
-    Attributes
-    ----------
-    model : torch model
-        Model to extract features from
-    outputs : list of torch tensors
-        List of outputs from hooks
-    features_per_layer : list of int
-        Number of features per hooked layer
-    module_list : list of lists
-        List of hooked layers, their names, and their indices in the model (if applicable)
-    """
-
-    def __init__(self, model_name='vgg16', model=None, use_cuda=None, param=None):
-
-        if model is not None:
-            assert model_name is None, 'model_name must be None if model is not None'
-            self.model = model
-        else:
-            if param is not None:
-                model_name = param.model_name
-
-            if model_name == 'vgg16':
-                self.model = models.vgg16(weights=models.VGG16_Weights.IMAGENET1K_V1)
-                # self.transform =  models.VGG16_Weights.IMAGENET1K_V1.transforms()
-            elif model_name == 'efficient_netb0':
-                self.model = models.efficientnet_b0(weights=models.EfficientNet_B0_Weights.IMAGENET1K_V1)
-            elif model_name == 'single_layer_vgg16':
-                self.model = load_single_layer_vgg16(keep_rgb=False)
-            elif model_name == 'single_layer_vgg16_rgb':
-                self.model = load_single_layer_vgg16(keep_rgb=True)
-            elif model_name == 'dino_vits16':
-                self.model = torch.hub.load('facebookresearch/dino:main', 'dino_vits16')
-
-        self.device = get_device(use_cuda)
-        self.model = self.model.to(self.device)
-
-        self.outputs = []
-        self.features_per_layer = []
-        self.selected_layers = []
-
-        self.get_layer_dict()
-
-        if model_name == 'single_layer_vgg16':
-            self.register_hooks(list(self.module_dict.keys()))
-
-        if param is not None:
-            self.register_hooks(param.model_layers)
-
-    def __call__(self, tensor_image):
-        tensor_image_dev = tensor_image.to(self.device)
-        return self.model(tensor_image_dev)
-
-    def hook_normal(self, module, input, output):
-        self.outputs.append(output)
-
-    def hook_last(self, module, input, output):
-        self.outputs.append(output)
-        assert False
-
-    def register_hooks(self, selected_layers):  # , selected_layer_pos):
-
-        self.features_per_layer = []
-        self.selected_layers = selected_layers.copy()
-        for ind in range(len(selected_layers)):
-
-            self.features_per_layer.append(
-                self.module_dict[selected_layers[ind]].out_channels)
-
-            if ind == len(selected_layers) - 1:
-                self.module_dict[selected_layers[ind]].register_forward_hook(self.hook_last)
-            else:
-                self.module_dict[selected_layers[ind]].register_forward_hook(self.hook_normal)
-
-    def get_layer_dict(self):
-        """Create a flat list of all modules as well as a dictionary of modules with 
-        keys describing the layers."""
-
-        named_modules = list(self.model.named_modules())
-        self.named_modules = [n for n in named_modules if len(list(n[1].named_modules())) == 1]
-        self.module_id_dict = dict([(x[0] + ' ' + x[1].__str__(), x[0]) for x in self.named_modules])
-        self.module_dict = dict([(x[0] + ' ' + x[1].__str__(), x[1]) for x in self.named_modules])
-
-
-class Classifier():
-    """"
-    Class to segment images. Contains both the NN model needed to extract features
-    and the classifier to predict the class of each pixel. By default loads the
-    single_layer_vgg16 model and the classifier is None.
-     
-    Parameters
-    ----------
-    model_path : str, optional
-        Path to RF model saved as joblib. Expects a parameters file in the same
-        location
-    
-    Attributes
-    ----------
-    random_forest : sklearn RF classifier
-        Classifier to predict the class of each pixel
-    param : Param
-        Parameters for model
-    model : Hookmodel
-        Model to extract features from the image
-
-    """
-
-    def __init__(self, model_path=None):
-
-        self.random_forest = None
-        self.param = None
-        self.model = None
-
-        if model_path is not None:
-            self.load_model(model_path)
-
-        else:
-            self.default_model()
-
-    def load_model(self, model_path):
-        """Load a pretrained model by loading the joblib model
-        and recreating the NN from the param file."""
-        
-        self.random_forest, self.param = load_trained_classifier(model_path)
-        self.model = Hookmodel(param=self.param)
-
-    def default_model(self):
-        """Set default model to single_layer_vgg16."""
-            
-        self.model = Hookmodel(model_name='single_layer_vgg16')
-        self.random_forest = None
-        self.param = Param(
-            model_name='single_layer_vgg16',
-            model_layers=list(self.model.module_dict.keys()),
-            scalings=[1,2],
-            order=1,
-            use_min_features=False,
-            normalize=True,
-        )
-
-    def save_classifier(self, save_path):
-        """Save the classifier to a joblib file and the parameters to a yaml file.
-        
-        Parameters
-        ----------
-        save_path : str
-            Path to save files to
-        """
-
-        dump(self.random_forest, save_path)
-        self.param.random_forest = save_path
-        self.param.save_parameters(Path(save_path).parent.joinpath('convpaint_params.yml'))
-
-
-    def segment_image_stack(self, image, save_path=None):
-        """Segment an image stack using a pretrained model. If save_path is not
-        None, save the zarr file to this path. Otherwise, return numpy array
-        
-        Parameters
-        ----------
-        image : np.ndarray
-            2D or 3D image stack to segment, with dimensions n,y,x or y,x
-        save_path : str
-            Path to save zarr file to
-
-        Returns
-        -------
-        np.ndarray
-            Segmented image stack. Either 2D (single image) or 3D with n,y,x
-            where n is the number of images in the stack.
-        """
-
-        if not ((image.ndim == 2) | (image.ndim == 3)):
-            raise Exception(f'Image must be 2D or 3D, not {image.ndim}')
-        single_image=False
-        if image.ndim == 2:
-            single_image = True
-            image = np.expand_dims(image, axis=0)
-        chunks = (1, image.shape[1], image.shape[2])
-
-        if save_path is not None:
-            im_out = zarr.open(save_path, mode='w', shape=image.shape,
-                               chunks=chunks, dtype=np.uint8)
-        else:
-            im_out = np.zeros(image.shape, dtype=np.uint8)
-
-        if self.param.normalize:
-            image_mean, image_std = compute_image_stats(image)
-            image = normalize_image(image, image_mean, image_std)
-
-        for i in range(image.shape[0]):
-            im_out[i] = predict_image(
-                image=image[i],
-                model=self.model,
-                classifier=self.random_forest,
-                scalings=self.param.scalings,
-                order=self.param.order,
-                use_min_features=self.param.use_min_features,
-                image_downsample=self.param.image_downsample)
-
-        if save_path is None:
-            if single_image:
-                return im_out[0]
-            else:
-                return im_out
